# public/openapi.yaml
openapi: 3.1.0
info:
  title: Azshrtr API
  version: 1.0.0
  summary: URL shortener and QR platform API
  description: 'Two authentication modes and clear tag groups in the explorer.


    **1. Product API (Bearer API key)** — tags under *Product API*

    `Authorization: Bearer az_live_…` or `az_test_…`

    Org context always comes from the key (never a path you can swap to another workspace).


    **2. Console session (cookie)** — Auth + Workspace + Links tags

    Sanctum cookie after login. Mutations need CSRF (`X-XSRF-TOKEN` / `X-CSRF-TOKEN`).

    API keys do **not** work on these routes.


    **Roles (console):** Owner manages billing; Owner/Admin manage members & API keys; Members can work
    with links/QR/domains/analytics.


    Interactive explorer: `/docs/api-explorer`. Spec: `/openapi.yaml`.

    '
  contact:
    name: Azodik Consulting Private Limited
    url: https://azodik.com
  license:
    name: MIT
    identifier: MIT
servers:
- url: https://azshrtr.test
  description: Local (Herd) — https://azshrtr.test
- url: http://localhost:8080
  description: Local (Docker Compose) — http://localhost:8080
- url: https://azshrtr.com
  description: Production — https://azshrtr.com
- url: '{baseUrl}'
  description: Custom deployed instance
  variables:
    baseUrl:
      default: https://azshrtr.com
      description: Application root URL (no trailing slash)
security: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: az_live_|az_test_
      description: 'Organization API key for the Product API (`/api/v1/me`, `/api/v1/links`, …). Send
        as `Authorization: Bearer az_live_…` or `az_test_…`. Scopes: links:read, links:write, qr:write,
        domains:read, analytics:read.'
    ConsoleSession:
      type: apiKey
      in: cookie
      name: azshrtr_session
      description: 'Laravel Sanctum SPA/session cookie used by the React console after login.

        Not an API key. Pair with CSRF (`X-XSRF-TOKEN` / `X-CSRF-TOKEN`) on unsafe methods.

        Required for `/api/v1/auth/me`, org console routes, billing, members, etc.

        '
...
paths:
...
  /api/v1/invitations/accept:
    post:
      tags:
      - Members
      operationId: acceptInvitation
      summary: Accept a member invitation
      security:
      - ConsoleSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - token
              properties:
                token:
                  type: string
      responses:
        200:
          description: Joined
      description: Accept a member invitation. Tagged under **Members**. See request/response schemas
        below.
...